requirements.txt
opencv-python-headless==4.5.4.60
torch==1.11.0
torchvision==0.12.0
numpy==1.21.6
aiogram==2.21
python-dotenv==0.19.0
